Init
This initial commit lays the foundation for N-ary Notebooks!  An open-source collection of interactive Jupyter notebooks dedicated to advancing Zero-Knowledge Proofs (ZKPs) and cryptographic protocol development. The project is designed for researchers, cryptography professionals, blockchain developers, and advanced students who want to collaborate on both the theory and practice of ZKPs.